+++ Packages/ai.fxn.fxn3d/Runtime/API/DotNetClient.cs
@@ -7,7 +7,6 @@
 
 namespace Function.API {
 
-    using System;
     using System.Collections.Generic;
     using System.IO;
     using System.Net.Http;
@@ -51,11 +50,10 @@ namespace Function.API {
         public override async Task<T?> Request<T> (
             string method,
             string path,
-            object? payload = default,
+            Dictionary<string, object?>? payload = default,
             Dictionary<string, string>? headers = default
         ) where T : class {
-            path = path.TrimStart('/');
-            using var message = new HttpRequestMessage(new HttpMethod(method), $"{url}/{path}");       
+            using var message = new HttpRequestMessage(new HttpMethod(method), $"{url}{path}");       
             // Populate headers
             if (!string.IsNullOrEmpty(accessKey))
                 message.Headers.Authorization = new AuthenticationHeaderValue(@"Bearer", accessKey);
@@ -75,28 +73,10 @@ namespace Function.API {
             if ((int)response.StatusCode >= 400) {
                 var errorPayload = JsonConvert.DeserializeObject<ErrorResponse>(responseStr);
                 var error = errorPayload?.errors?[0]?.message ?? @"An unknown error occurred";
-                throw new InvalidOperationException(error);
+                throw new FunctionAPIException(error, (int)response.StatusCode);
             }
             // Return
             return JsonConvert.DeserializeObject<T>(responseStr)!;
-        }
-
-        /// <summary>
-        /// Query the Function graph API.
-        /// </summary>
-        /// <param name="query">Graph query.</param>
-        /// <param name="key">Query result key.</param>
-        /// <param name="input">Query inputs.</param>
-        public override async Task<T?> Query<T> (
-            string query,
-            Dictionary<string, object?>? variables = default
-        ) where T : class {
-            var response = await Request<GraphResponse<T>>(
-                @"POST",
-                @"/graph",
-                new GraphRequest { query = query, variables = variables }
-            );
-            return response!.data;
         }
 
         /// <summary>