+++ Changelog.md
@@ -1,4 +1,6 @@
 ## 0.0.30
++ Added `FunctionAPIException` exception class for explicitly catching API errors.
++ Removed `Predictor.predictions` field.
 + Function now requires Unity 2022.3+
 
 ## 0.0.29